refactor(export-image): dry out drawEarth
